+++ software_esp32/src/VdmNet.cpp
@@ -308,3 +308,15 @@ void CVdmNet::checkNet()
   }
 }
 
+void CVdmNet::checkWifi()
+{
+  uint8_t WifiStatus = WiFi.status();
+  UART_DBG.println("Wifi status "+ String(WifiStatus));
+  if (wifiState==wifiConnected) {
+      if (WifiStatus!= WL_CONNECTED) {
+      WiFi.disconnect();
+      WiFi.reconnect();
+      UART_DBG.println("Wifi reconnect");
+    }
+  }
+}