+++ software_esp32/src/mqtt.cpp
@@ -112,6 +112,7 @@ CMqtt::CMqtt()
 uint8_t CMqtt::checkForPublish() 
 {
     uint8_t result=publishNothing;
+    // if (!topicsReceived) return result;
     uint8_t i;
 
     forcePublish = true;
@@ -129,7 +130,9 @@ uint8_t CMqtt::checkForPublish()
             if (VdmSystem.systemMessage.length()>0) result = publishCommon; 
             if (lastCommonValues.heatControl!=VdmConfig.configFlash.valvesControlConfig.heatControl) result = publishCommon; 
             if (lastCommonValues.parkingPosition!=VdmConfig.configFlash.valvesControlConfig.parkingPosition) result = publishCommon; 
-            if (lastCommonValues.systemState!=VdmSystem.systemState) result = publishCommon; 
+            if (lastCommonValues.systemState!=VdmSystem.systemState) result = publishCommon;
+            String upTime = VdmSystem.getUpTime();
+            if (lastCommonValues.upTime!=upTime) result = publishCommon;
         }
         // valves
         for (i=0;i<ACTUATOR_COUNT;i++) {
@@ -219,6 +222,7 @@ void CMqtt::disconnect() {
 
 void CMqtt::reconnect() 
 {
+    topicsReceived=false;
     char topicstr[MAINTOPIC_LEN+50];
     char nrstr[11];
     char* mqttUser = NULL;
@@ -260,10 +264,18 @@ void CMqtt::reconnect()
     memset(topicstr,0x0,sizeof(topicstr));
     strncat(topicstr,mqtt_commonTopic,sizeof(topicstr) - strlen (topicstr) - 1);
     len = strlen(topicstr);
-    strncat(topicstr, "heatControl",sizeof(topicstr) - strlen (topicstr) - 1);
+    if (VdmConfig.configFlash.protConfig.protocolFlags.publishSeparate) {
+        strncat(topicstr, "setHeatControl",sizeof(topicstr) - strlen (topicstr) - 1);
+    } else {
+        strncat(topicstr, "heatControl",sizeof(topicstr) - strlen (topicstr) - 1);
+    }
     mqtt_client.subscribe(topicstr);    
     topicstr[len] = '\0';
-    strncat(topicstr, "parkPosition",sizeof(topicstr) - strlen (topicstr) - 1);
+    if (VdmConfig.configFlash.protConfig.protocolFlags.publishSeparate) {
+        strncat(topicstr, "setParkPosition",sizeof(topicstr) - strlen (topicstr) - 1);
+    } else {
+        strncat(topicstr, "parkPosition",sizeof(topicstr) - strlen (topicstr) - 1);
+    }
     mqtt_client.subscribe(topicstr);    
 
     for (uint8_t x = 0;x<ACTUATOR_COUNT;x++) {
@@ -301,17 +313,22 @@ void CMqtt::reconnect()
                     } 
                 }
                 // window state
-                if (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[x].controlFlags.windowInstalled) {
+                if (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[x].link==0) {
+                    if (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[x].controlFlags.windowInstalled) {
+                        topicstr[len] = '\0';
+                        strncat(topicstr, "/window/state",sizeof(topicstr) - strlen (topicstr) - 1);
+                        mqtt_client.subscribe(topicstr);  
+                        // window target (%)
+                        topicstr[len] = '\0';
+                        strncat(topicstr, "/window/target",sizeof(topicstr) - strlen (topicstr) - 1);
+                        mqtt_client.subscribe(topicstr);  
+                    }
+                }
+                 // control active
+                if (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[x].link==0) {
                     topicstr[len] = '\0';
-                    strncat(topicstr, "/window/state",sizeof(topicstr) - strlen (topicstr) - 1);
-                    mqtt_client.subscribe(topicstr);  
-                    // window target (%)
-                    topicstr[len] = '\0';
-                    strncat(topicstr, "/window/target",sizeof(topicstr) - strlen (topicstr) - 1);
-                    mqtt_client.subscribe(topicstr);  
-                    #ifdef EnvDevelop
-                        UART_DBG.println("MQTT : subscribe window "+String(VdmConfig.configFlash.valvesConfig.valveConfig[x].name));
-                    #endif
+                    strncat(topicstr, "/control/mode",sizeof(topicstr) - strlen (topicstr) - 1);
+                    mqtt_client.subscribe(topicstr);                 
                 }
             }
         }
@@ -334,6 +351,7 @@ void CMqtt::callback(char* topic, byte* payload, unsigned int length)
     uint8_t i;
     uint8_t idx;
     uint8_t val8;
+    bool topicFound = false;
 
     if (VdmConfig.configFlash.netConfig.syslogLevel>=VISMODE_DETAIL) {
                syslog.log(LOG_DEBUG, "MQTT: callback "+String(topic));
@@ -353,14 +371,32 @@ void CMqtt::callback(char* topic, byte* payload, unsigned int length)
             memset(item,0x0,sizeof(item));
             pt= (char*) topic;
             pt+= strlen(mqtt_commonTopic);
-            if (strncmp(pt,"heatControl",sizeof("heatControl"))==0) {
-                val8=atoi(value);
+            if (VdmConfig.configFlash.protConfig.protocolFlags.publishSeparate) {
+                topicFound = (strncmp(pt,"setHeatControl",sizeof("setHeatControl"))==0);
+            } else {
+                topicFound = (strncmp(pt,"heatControl",sizeof("heatControl"))==0);
+            }
+            if (topicFound) {
+                val8=VdmConfig.configFlash.valvesControlConfig.heatControl;
+                if (isFloat(value)) {
+                    val8=atoi(value);
+                } else {
+                    if (strncmp(value,"manual",sizeof("manual"))==0) val8=0;
+                    if (strncmp(value,"heat",sizeof("heat"))==0) val8=1;
+                    if (strncmp(value,"cool",sizeof("cool"))==0) val8=2;
+                    if (strncmp(value,"off",sizeof("off"))==0) val8=3;
+                }
                 if (val8!=VdmConfig.configFlash.valvesControlConfig.heatControl) {
                     VdmConfig.configFlash.valvesControlConfig.heatControl=val8;
                     VdmConfig.writeValvesControlConfig(false,false); 
                 }
             } 
-            if (strncmp(pt,"parkPosition",sizeof("parkPosition"))==0) {
+            if (VdmConfig.configFlash.protConfig.protocolFlags.publishSeparate) {
+                topicFound = (strncmp(pt,"setParkPosition",sizeof("setParkPosition"))==0);
+            } else {
+                topicFound = (strncmp(pt,"parkPosition",sizeof("parkPosition"))==0);
+            }
+            if (topicFound) {
                 val8=atoi(value); 
                 if (val8!=VdmConfig.configFlash.valvesControlConfig.heatControl) {
                     VdmConfig.configFlash.valvesControlConfig.parkingPosition=val8;
@@ -427,13 +463,12 @@ void CMqtt::callback(char* topic, byte* payload, unsigned int length)
                             if (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[idx].targetSource==0)
                                 PiControl[idx].dynOffset=atoi(value);
                         }
-                    } else if (strncmp(pt,"/window/state",13)==0) {
-                        if (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[idx].controlFlags.active) {
-                                PiControl[idx].setWindowAction(atoi(value));
-                        }
-                    } else if (strncmp(pt,"/window/target",14)==0) {
-                        if (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[idx].controlFlags.active) {
-                                PiControl[idx].windowOpenTarget=atoi(value);
+                    } 
+                    else if (strncmp(pt,"/window/target",14)==0) {
+                        if (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[idx].link==0) {
+                            if (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[idx].controlFlags.active) {
+                                    PiControl[idx].windowOpenTarget=atoi(value);
+                            }
                         }
                     }
                     if (VdmConfig.configFlash.netConfig.syslogLevel>=VISMODE_DETAIL) {
@@ -444,22 +479,48 @@ void CMqtt::callback(char* topic, byte* payload, unsigned int length)
                         syslog.log(LOG_DEBUG, "MQTT: found topic, but not a number "+String(item)+String(pt)+" : "+String(value));
                     }  
                 }
+                // handling for plain text
+                if (strncmp(pt,"/control/mode",13)==0) {
+                    if (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[idx].link==0) {
+                        if (isFloat(value)) {
+                            val8=atoi(value);
+                            PiControl[idx].controlActive=bool(val8);
+                        } else {
+                            if (strncmp(value,"off",3)==0)  PiControl[idx].setControlActive(0); //PiControl[idx].controlActive=0;
+                            if (strncmp(value,"auto",4)==0)  PiControl[idx].setControlActive(1); //controlActive=1;
+                        } 
+                        if (PiControl[idx].controlActive==0) PiControl[idx].setPosition(0);    
+                    }
+                }
+                if (strncmp(pt,"/window/state",13)==0) {
+                    if (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[idx].link==0) {
+                        if (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[idx].controlFlags.active) {
+                            if (isFloat(value)) {
+                                PiControl[idx].setWindowAction(atoi(value));
+                            } else {
+                                if (strncmp(value,"close",5)==0)  PiControl[idx].setWindowAction(0);
+                                if (strncmp(value,"open",4)==0)  PiControl[idx].setWindowAction(1);
+                            }       
+                        }
+                    }
+                }
             } else {
                 if (VdmConfig.configFlash.netConfig.syslogLevel>=VISMODE_DETAIL) {
                     syslog.log(LOG_DEBUG, "MQTT: not found topic "+String(item));
                 }   
             }
-            
         }
     }
+    if (!topicsReceived && found) topicsReceived=true; 
 }
 
 void CMqtt::publish_valves () {
     char topicstr[MAINTOPIC_LEN+30];
     char nrstr[11];
-    char valstr[10];
+    char valstr[50];
     int8_t tempIdx;
     uint8_t len;
+    const char valveStates[9][11] =  {"","idle","opens","closes","blocked","unknown","no valve","full open","connected"};
 
     for (uint8_t x = 0;x<ACTUATOR_COUNT;x++) {
         if (VdmConfig.configFlash.valvesConfig.valveConfig[x].active) {
@@ -482,7 +543,7 @@ void CMqtt::publish_valves () {
                     lastValveValues[x].position=StmApp.actuators[x].actual_position;
                 }
                 // target
-                if (VdmConfig.configFlash.protConfig.protocolFlags.publishTarget)  {
+                if (!VdmConfig.configFlash.protConfig.protocolFlags.publishSeparate)  {
                     if ((lastValveValues[x].target!=StmApp.actuators[x].target_position) || forcePublish || lastValveValues[x].publishTimeOut) {
                         topicstr[len] = '\0';
                         strncat(topicstr, "/target",sizeof(topicstr) - strlen (topicstr) - 1);
@@ -495,8 +556,26 @@ void CMqtt::publish_valves () {
                 if ((lastValveValues[x].state!=StmApp.actuators[x].state) || forcePublish || lastValveValues[x].publishTimeOut) {
                     topicstr[len] = '\0';
                     strncat(topicstr, "/state",sizeof(topicstr) - strlen (topicstr) - 1);
-                    itoa(StmApp.actuators[x].state, valstr, 10);
+                    if (VdmConfig.configFlash.protConfig.protocolFlags.publishPlainText) {
+                        if (StmApp.actuators[x].state<9) {
+                            strcpy(valstr,valveStates[StmApp.actuators[x].state]);
+                        } else strcpy(valstr,"");
+                    } else {
+                        itoa(StmApp.actuators[x].state, valstr, 10);        
+                    }
                     mqtt_client.publish(topicstr, valstr,VdmConfig.configFlash.protConfig.protocolFlags.publishRetained);
+                    
+                    if (StmApp.actuators[x].calibration) {
+                        topicstr[len] = '\0';
+                        strncat(topicstr, "/calibration",sizeof(topicstr) - strlen (topicstr) - 1);
+                        mqtt_client.publish(topicstr, VdmSystem.localTime().c_str(),VdmConfig.configFlash.protConfig.protocolFlags.publishRetained);
+                    }
+                    if ((VdmConfig.configFlash.valvesControlConfig.valveControlConfig[x].link>0) && (VdmConfig.configFlash.valvesControlConfig.valveControlConfig[x].link<13)) {
+                        topicstr[len] = '\0';
+                        strncat(topicstr, "/control/link",sizeof(topicstr) - strlen (topicstr) - 1);
+                        strcpy(valstr,VdmConfig.configFlash.valvesConfig.valveConfig[VdmConfig.configFlash.valvesControlConfig.valveControlConfig[x].link-1].name);
+                        mqtt_client.publish(topicstr, valstr,VdmConfig.configFlash.protConfig.protocolFlags.publishRetained);
+                    }
                     lastValveValues[x].state=StmApp.actuators[x].state;
                 }
                 // meancurrent
@@ -527,6 +606,7 @@ void CMqtt::publish_valves () {
                         lastValveValues[x].temp2=StmApp.actuators[x].temp2;
                     }
                 }
+         
                 lastValveValues[x].publishNow=false;
                 lastValveValues[x].publishTimeOut=false;
                 lastValveValues[x].ts=millis();
@@ -579,16 +659,24 @@ void CMqtt::publish_common ()
 {
     char topicstr[MAINTOPIC_LEN+30];
     char nrstr[11];
-    char valstr[10];
+    char valstr[20];
     int8_t tempIdx;
     uint8_t len;
-    
+    const char heatControlStates[4][11] =  {"manual","heat","cool","off"};
+    const char systemStates[3][11] =  {"ok","info","error"};
+
     memset(topicstr,0x0,sizeof(topicstr));
     strncat(topicstr,mqtt_commonTopic,sizeof(topicstr) - strlen (topicstr) - 1);
     len = strlen(topicstr);
     if ((!VdmConfig.configFlash.protConfig.protocolFlags.publishOnChange) || (lastCommonValues.heatControl!=VdmConfig.configFlash.valvesControlConfig.heatControl)) {
         strncat(topicstr, "heatControl",sizeof(topicstr) - strlen (topicstr) - 1);
-        itoa(VdmConfig.configFlash.valvesControlConfig.heatControl, valstr, 10);        
+        if (VdmConfig.configFlash.protConfig.protocolFlags.publishPlainText) {
+            if (VdmConfig.configFlash.valvesControlConfig.heatControl<4) {
+                strcpy(valstr,heatControlStates[VdmConfig.configFlash.valvesControlConfig.heatControl]);
+            } else strcpy(valstr,"");
+        } else {
+            itoa(VdmConfig.configFlash.valvesControlConfig.heatControl, valstr, 10);        
+        }
         mqtt_client.publish(topicstr, valstr,VdmConfig.configFlash.protConfig.protocolFlags.publishRetained); 
         lastCommonValues.heatControl=VdmConfig.configFlash.valvesControlConfig.heatControl; 
     }
@@ -604,7 +692,13 @@ void CMqtt::publish_common ()
     
     if ((!VdmConfig.configFlash.protConfig.protocolFlags.publishOnChange) || (lastCommonValues.systemState!=VdmSystem.systemState)) {
         strncat(topicstr, "state",sizeof(topicstr) - strlen (topicstr) - 1);
-        itoa(VdmSystem.systemState , valstr, 10);        
+        if (VdmConfig.configFlash.protConfig.protocolFlags.publishPlainText) {
+            if (VdmSystem.systemState<3) {
+                strcpy(valstr,systemStates[VdmSystem.systemState]);
+            } else strcpy(valstr,"");
+        } else {
+            itoa(VdmSystem.systemState , valstr, 10);        
+        }
         mqtt_client.publish(topicstr, valstr,VdmConfig.configFlash.protConfig.protocolFlags.publishRetained);  
         lastCommonValues.systemState=VdmSystem.systemState;
     }
@@ -612,7 +706,10 @@ void CMqtt::publish_common ()
     if (VdmConfig.configFlash.protConfig.protocolFlags.publishUpTime) {
         strncat(topicstr, "uptime",sizeof(topicstr) - strlen (topicstr) - 1);
         String upTime = VdmSystem.getUpTime();
-        mqtt_client.publish(topicstr, (const char*) (upTime.c_str()),VdmConfig.configFlash.protConfig.protocolFlags.publishRetained); 
+        if ((!VdmConfig.configFlash.protConfig.protocolFlags.publishOnChange) || (lastCommonValues.upTime!=upTime)) {
+            lastCommonValues.upTime = upTime;
+            mqtt_client.publish(topicstr, (const char*) (upTime.c_str()),VdmConfig.configFlash.protConfig.protocolFlags.publishRetained); 
+        }
     }
     
     topicstr[len] = '\0';